@@ -1,14 +1,14 @@
 import '/auth/firebase_auth/auth_util.dart';
-import '/backend/backend.dart';
+import '/backend/firebase_storage/storage.dart';
 import '/flutter_flow/flutter_flow_animations.dart';
 import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import '/flutter_flow/flutter_flow_widgets.dart';
+import '/flutter_flow/upload_data.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_animate/flutter_animate.dart';
-import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
@@ -16,7 +16,7 @@ import 'login_page_model.dart';
 export 'login_page_model.dart';
 
 class LoginPageWidget extends StatefulWidget {
-  const LoginPageWidget({Key? key}) : super(key: key);
+  const LoginPageWidget({super.key});
 
   @override
   _LoginPageWidgetState createState() => _LoginPageWidgetState();
@@ -44,8 +44,8 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
           curve: Curves.easeInOut,
           delay: 100.ms,
           duration: 600.ms,
-          begin: Offset(0.0, 20.0),
-          end: Offset(0.0, 0.0),
+          begin: const Offset(0.0, 20.0),
+          end: const Offset(0.0, 0.0),
         ),
       ],
     ),
@@ -64,8 +64,8 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
           curve: Curves.easeInOut,
           delay: 0.ms,
           duration: 600.ms,
-          begin: Offset(0.0, 60.0),
-          end: Offset(0.0, 0.0),
+          begin: const Offset(0.0, 60.0),
+          end: const Offset(0.0, 0.0),
         ),
       ],
     ),
@@ -84,8 +84,8 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
           curve: Curves.easeInOut,
           delay: 200.ms,
           duration: 600.ms,
-          begin: Offset(0.0, 30.0),
-          end: Offset(0.0, 0.0),
+          begin: const Offset(0.0, 30.0),
+          end: const Offset(0.0, 0.0),
         ),
       ],
     ),
@@ -102,12 +102,23 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
       initialIndex: 0,
     )..addListener(() => setState(() {}));
     _model.logemailAddressController ??= TextEditingController();
+    _model.logemailAddressFocusNode ??= FocusNode();
+
     _model.logpasswordController ??= TextEditingController();
-    _model.signimageController ??= TextEditingController();
+    _model.logpasswordFocusNode ??= FocusNode();
+
     _model.signemailaddController ??= TextEditingController();
+    _model.signemailaddFocusNode ??= FocusNode();
+
     _model.signnameController ??= TextEditingController();
+    _model.signnameFocusNode ??= FocusNode();
+
     _model.signpassController ??= TextEditingController();
+    _model.signpassFocusNode ??= FocusNode();
+
     _model.signpassCheckController ??= TextEditingController();
+    _model.signpassCheckFocusNode ??= FocusNode();
+
     setupAnimations(
       animationsMap.values.where((anim) =>
           anim.trigger == AnimationTrigger.onActionTrigger ||
@@ -127,60 +138,73 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
 
   @override
   Widget build(BuildContext context) {
+    if (isiOS) {
+      SystemChrome.setSystemUIOverlayStyle(
+        SystemUiOverlayStyle(
+          statusBarBrightness: Theme.of(context).brightness,
+          systemStatusBarContrastEnforced: true,
+        ),
+      );
+    }
+
     context.watch<FFAppState>();
 
     return GestureDetector(
-      onTap: () => FocusScope.of(context).requestFocus(_model.unfocusNode),
+      onTap: () => _model.unfocusNode.canRequestFocus
+          ? FocusScope.of(context).requestFocus(_model.unfocusNode)
+          : FocusScope.of(context).unfocus(),
       child: Scaffold(
         key: scaffoldKey,
         backgroundColor: Colors.white,
-        body: Container(
+        body: SizedBox(
           width: double.infinity,
           height: double.infinity,
           child: Stack(
-            alignment: AlignmentDirectional(0.0, 1.0),
+            alignment: const AlignmentDirectional(0.0, 1.0),
             children: [
               Align(
-                alignment: AlignmentDirectional(0.00, 1.00),
+                alignment: const AlignmentDirectional(0.00, 1.00),
                 child: Container(
                   width: double.infinity,
                   height: double.infinity,
-                  decoration: BoxDecoration(
+                  decoration: const BoxDecoration(
                     color: Colors.white,
                   ),
-                  alignment: AlignmentDirectional(0.00, 1.00),
+                  alignment: const AlignmentDirectional(0.00, 1.00),
                   child: Align(
-                    alignment: AlignmentDirectional(0.00, 1.00),
+                    alignment: const AlignmentDirectional(0.00, 1.00),
                     child: Column(
                       mainAxisSize: MainAxisSize.max,
                       mainAxisAlignment: MainAxisAlignment.end,
                       children: [
-                        Padding(
-                          padding: EdgeInsetsDirectional.fromSTEB(
-                              0.0, 50.0, 0.0, 0.0),
-                          child: Row(
-                            mainAxisSize: MainAxisSize.max,
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: [
-                              Stack(
-                                children: [
-                                  Image.asset(
-                                    'assets/images/logo1.png',
-                                    width: 150.0,
-                                    fit: BoxFit.fitWidth,
+                        Row(
+                          mainAxisSize: MainAxisSize.max,
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            Stack(
+                              children: [
+                                Align(
+                                  alignment: const AlignmentDirectional(0.00, 0.00),
+                                  child: ClipRRect(
+                                    borderRadius: BorderRadius.circular(15.0),
+                                    child: Image.asset(
+                                      'assets/images/IMG_0018-removebg-preview.png',
+                                      width: 150.0,
+                                      fit: BoxFit.fitWidth,
+                                    ),
                                   ),
-                                ],
-                              ),
-                            ],
-                          ).animateOnPageLoad(
-                              animationsMap['rowOnPageLoadAnimation']!),
-                        ),
+                                ),
+                              ],
+                            ),
+                          ],
+                        ).animateOnPageLoad(
+                            animationsMap['rowOnPageLoadAnimation']!),
                         Expanded(
                           child: Align(
-                            alignment: AlignmentDirectional(0.00, 1.00),
+                            alignment: const AlignmentDirectional(0.00, 1.00),
                             child: Container(
                               width: double.infinity,
-                              constraints: BoxConstraints(
+                              constraints: const BoxConstraints(
                                 maxWidth: 570.0,
                               ),
                               decoration: BoxDecoration(
@@ -194,26 +218,26 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                     child: Column(
                                       children: [
                                         Align(
-                                          alignment: Alignment(0.0, 0),
+                                          alignment: const Alignment(0.0, 0),
                                           child: TabBar(
                                             isScrollable: true,
                                             labelColor:
                                                 FlutterFlowTheme.of(context)
-                                                    .mainC,
+                                                    .main1,
                                             unselectedLabelColor:
-                                                Color(0xFFB6BCBF),
+                                                const Color(0xFFB6BCBF),
                                             labelPadding:
-                                                EdgeInsetsDirectional.fromSTEB(
+                                                const EdgeInsetsDirectional.fromSTEB(
                                                     24.0, 0.0, 24.0, 0.0),
                                             labelStyle:
                                                 FlutterFlowTheme.of(context)
                                                     .titleMedium,
-                                            unselectedLabelStyle: TextStyle(),
+                                            unselectedLabelStyle: const TextStyle(),
                                             indicatorColor:
                                                 FlutterFlowTheme.of(context)
                                                     .main1,
                                             indicatorWeight: 2.0,
-                                            tabs: [
+                                            tabs: const [
                                               Tab(
                                                 text: '로그인',
                                               ),
@@ -235,7 +259,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                   children: [
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   24.0,
                                                                   20.0,
@@ -244,6 +268,8 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                       child: TextFormField(
                                                         controller: _model
                                                             .logemailAddressController,
+                                                        focusNode: _model
+                                                            .logemailAddressFocusNode,
                                                         obscureText: false,
                                                         decoration:
                                                             InputDecoration(
@@ -270,7 +296,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           enabledBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0xFFD1D1D1),
                                                               width: 1.0,
@@ -283,7 +309,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           focusedBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0x00000000),
                                                               width: 1.0,
@@ -326,7 +352,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                   .of(context)
                                                               .primaryBackground,
                                                           contentPadding:
-                                                              EdgeInsetsDirectional
+                                                              const EdgeInsetsDirectional
                                                                   .fromSTEB(
                                                                       20.0,
                                                                       24.0,
@@ -357,7 +383,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   24.0,
                                                                   12.0,
@@ -366,6 +392,8 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                       child: TextFormField(
                                                         controller: _model
                                                             .logpasswordController,
+                                                        focusNode: _model
+                                                            .logpasswordFocusNode,
                                                         obscureText: !_model
                                                             .logpasswordVisibility,
                                                         decoration:
@@ -393,7 +421,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           enabledBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0xFFD1D1D1),
                                                               width: 1.0,
@@ -406,7 +434,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           focusedBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0x00000000),
                                                               width: 1.0,
@@ -449,7 +477,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                   .of(context)
                                                               .primaryBackground,
                                                           contentPadding:
-                                                              EdgeInsetsDirectional
+                                                              const EdgeInsetsDirectional
                                                                   .fromSTEB(
                                                                       20.0,
                                                                       24.0,
@@ -503,7 +531,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   24.0,
                                                                   16.0,
@@ -538,14 +566,14 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                               width: 140.0,
                                                               height: 45.0,
                                                               padding:
-                                                                  EdgeInsetsDirectional
+                                                                  const EdgeInsetsDirectional
                                                                       .fromSTEB(
                                                                           0.0,
                                                                           0.0,
                                                                           0.0,
                                                                           0.0),
                                                               iconPadding:
-                                                                  EdgeInsetsDirectional
+                                                                  const EdgeInsetsDirectional
                                                                       .fromSTEB(
                                                                           0.0,
                                                                           0.0,
@@ -616,14 +644,14 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                               width: 130.0,
                                                               height: 45.0,
                                                               padding:
-                                                                  EdgeInsetsDirectional
+                                                                  const EdgeInsetsDirectional
                                                                       .fromSTEB(
                                                                           0.0,
                                                                           0.0,
                                                                           0.0,
                                                                           0.0),
                                                               iconPadding:
-                                                                  EdgeInsetsDirectional
+                                                                  const EdgeInsetsDirectional
                                                                       .fromSTEB(
                                                                           0.0,
                                                                           0.0,
@@ -631,7 +659,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                           0.0),
                                                               color: FlutterFlowTheme
                                                                       .of(context)
-                                                                  .mainC,
+                                                                  .main1,
                                                               textStyle:
                                                                   FlutterFlowTheme.of(
                                                                           context)
@@ -646,7 +674,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                       ),
                                                               elevation: 3.0,
                                                               borderSide:
-                                                                  BorderSide(
+                                                                  const BorderSide(
                                                                 color: Colors
                                                                     .transparent,
                                                                 width: 1.0,
@@ -662,7 +690,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   20.0,
                                                                   0.0,
@@ -677,7 +705,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                         children: [
                                                           Padding(
                                                             padding:
-                                                                EdgeInsetsDirectional
+                                                                const EdgeInsetsDirectional
                                                                     .fromSTEB(
                                                                         0.0,
                                                                         20.0,
@@ -692,7 +720,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                     fontFamily:
                                                                         FlutterFlowTheme.of(context)
                                                                             .bodySmallFamily,
-                                                                    color: Color(
+                                                                    color: const Color(
                                                                         0xC5000000),
                                                                     fontSize:
                                                                         14.0,
@@ -708,7 +736,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   0.0,
                                                                   4.0,
@@ -723,7 +751,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                         children: [
                                                           Padding(
                                                             padding:
-                                                                EdgeInsetsDirectional
+                                                                const EdgeInsetsDirectional
                                                                     .fromSTEB(
                                                                         8.0,
                                                                         8.0,
@@ -734,7 +762,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                               borderColor:
                                                                   FlutterFlowTheme.of(
                                                                           context)
-                                                                      .mainC,
+                                                                      .main1,
                                                               borderRadius:
                                                                   12.0,
                                                               borderWidth: 1.0,
@@ -744,7 +772,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                     .google,
                                                                 color: FlutterFlowTheme.of(
                                                                         context)
-                                                                    .mainC,
+                                                                    .sub1,
                                                                 size: 16.0,
                                                               ),
                                                               onPressed: () {
@@ -755,7 +783,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           ),
                                                           Padding(
                                                             padding:
-                                                                EdgeInsetsDirectional
+                                                                const EdgeInsetsDirectional
                                                                     .fromSTEB(
                                                                         8.0,
                                                                         8.0,
@@ -766,17 +794,17 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                               borderColor:
                                                                   FlutterFlowTheme.of(
                                                                           context)
-                                                                      .mainC,
+                                                                      .main1,
                                                               borderRadius:
                                                                   12.0,
                                                               borderWidth: 1.0,
                                                               buttonSize: 44.0,
-                                                              icon: Icon(
-                                                                FFIcons
-                                                                    .knavericon,
+                                                              icon: FaIcon(
+                                                                FontAwesomeIcons
+                                                                    .accusoft,
                                                                 color: FlutterFlowTheme.of(
                                                                         context)
-                                                                    .mainC,
+                                                                    .sub1,
                                                                 size: 16.0,
                                                               ),
                                                               onPressed: () {
@@ -787,7 +815,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           ),
                                                           Padding(
                                                             padding:
-                                                                EdgeInsetsDirectional
+                                                                const EdgeInsetsDirectional
                                                                     .fromSTEB(
                                                                         8.0,
                                                                         8.0,
@@ -798,17 +826,17 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                               borderColor:
                                                                   FlutterFlowTheme.of(
                                                                           context)
-                                                                      .mainC,
+                                                                      .main1,
                                                               borderRadius:
                                                                   12.0,
                                                               borderWidth: 1.0,
                                                               buttonSize: 44.0,
-                                                              icon: Icon(
-                                                                FFIcons
-                                                                    .kinstalogo,
+                                                              icon: FaIcon(
+                                                                FontAwesomeIcons
+                                                                    .android,
                                                                 color: FlutterFlowTheme.of(
                                                                         context)
-                                                                    .mainC,
+                                                                    .sub1,
                                                                 size: 16.0,
                                                               ),
                                                               onPressed: () {
@@ -828,132 +856,265 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                   mainAxisSize:
                                                       MainAxisSize.max,
                                                   children: [
-                                                    Padding(
-                                                      padding:
-                                                          EdgeInsetsDirectional
-                                                              .fromSTEB(
-                                                                  24.0,
-                                                                  12.0,
-                                                                  24.0,
-                                                                  0.0),
-                                                      child: TextFormField(
-                                                        controller: _model
-                                                            .signimageController,
-                                                        obscureText: false,
-                                                        decoration:
-                                                            InputDecoration(
-                                                          labelText:
-                                                              '이미지 업데이트 ',
-                                                          labelStyle:
-                                                              FlutterFlowTheme.of(
-                                                                      context)
-                                                                  .bodySmall
+                                                    Column(
+                                                      mainAxisSize:
+                                                          MainAxisSize.max,
+                                                      children: [
+                                                        Align(
+                                                          alignment:
+                                                              const AlignmentDirectional(
+                                                                  -1.00, 0.00),
+                                                          child: Padding(
+                                                            padding:
+                                                                const EdgeInsetsDirectional
+                                                                    .fromSTEB(
+                                                                        25.0,
+                                                                        10.0,
+                                                                        0.0,
+                                                                        0.0),
+                                                            child: Text(
+                                                              'Upload a profile image *',
+                                                              style: FlutterFlowTheme
+                                                                      .of(context)
+                                                                  .titleSmall
                                                                   .override(
                                                                     fontFamily:
                                                                         FlutterFlowTheme.of(context)
-                                                                            .bodySmallFamily,
-                                                                    fontSize:
-                                                                        14.0,
+                                                                            .titleSmallFamily,
+                                                                    color: FlutterFlowTheme.of(
+                                                                            context)
+                                                                        .main1,
                                                                     useGoogleFonts: GoogleFonts
                                                                             .asMap()
                                                                         .containsKey(
-                                                                            FlutterFlowTheme.of(context).bodySmallFamily),
+                                                                            FlutterFlowTheme.of(context).titleSmallFamily),
                                                                   ),
-                                                          hintStyle:
-                                                              FlutterFlowTheme.of(
-                                                                      context)
-                                                                  .bodySmall,
-                                                          enabledBorder:
-                                                              OutlineInputBorder(
-                                                            borderSide:
-                                                                BorderSide(
-                                                              color: Color(
-                                                                  0xFFD1D1D1),
-                                                              width: 1.0,
                                                             ),
-                                                            borderRadius:
-                                                                BorderRadius
-                                                                    .circular(
-                                                                        8.0),
                                                           ),
-                                                          focusedBorder:
-                                                              OutlineInputBorder(
-                                                            borderSide:
-                                                                BorderSide(
-                                                              color: Color(
-                                                                  0x00000000),
-                                                              width: 1.0,
-                                                            ),
-                                                            borderRadius:
-                                                                BorderRadius
-                                                                    .circular(
-                                                                        8.0),
-                                                          ),
-                                                          errorBorder:
-                                                              OutlineInputBorder(
-                                                            borderSide:
-                                                                BorderSide(
-                                                              color: FlutterFlowTheme
-                                                                      .of(context)
-                                                                  .primary,
-                                                              width: 1.0,
-                                                            ),
-                                                            borderRadius:
-                                                                BorderRadius
-                                                                    .circular(
-                                                                        8.0),
-                                                          ),
-                                                          focusedErrorBorder:
-                                                              OutlineInputBorder(
-                                                            borderSide:
-                                                                BorderSide(
-                                                              color: FlutterFlowTheme
-                                                                      .of(context)
-                                                                  .primary,
-                                                              width: 1.0,
-                                                            ),
-                                                            borderRadius:
-                                                                BorderRadius
-                                                                    .circular(
-                                                                        8.0),
-                                                          ),
-                                                          filled: true,
-                                                          fillColor: FlutterFlowTheme
-                                                                  .of(context)
-                                                              .primaryBackground,
-                                                          contentPadding:
-                                                              EdgeInsetsDirectional
-                                                                  .fromSTEB(
-                                                                      20.0,
-                                                                      24.0,
-                                                                      20.0,
-                                                                      24.0),
                                                         ),
-                                                        style:
-                                                            FlutterFlowTheme.of(
-                                                                    context)
-                                                                .bodyMedium
-                                                                .override(
-                                                                  fontFamily: FlutterFlowTheme.of(
-                                                                          context)
-                                                                      .bodyMediumFamily,
-                                                                  fontSize:
-                                                                      14.0,
-                                                                  useGoogleFonts: GoogleFonts
-                                                                          .asMap()
-                                                                      .containsKey(
-                                                                          FlutterFlowTheme.of(context)
-                                                                              .bodyMediumFamily),
+                                                        Row(
+                                                          mainAxisSize:
+                                                              MainAxisSize.max,
+                                                          mainAxisAlignment:
+                                                              MainAxisAlignment
+                                                                  .spaceBetween,
+                                                          children: [
+                                                            Align(
+                                                              alignment:
+                                                                  const AlignmentDirectional(
+                                                                      -1.00,
+                                                                      0.00),
+                                                              child: Padding(
+                                                                padding:
+                                                                    const EdgeInsetsDirectional
+                                                                        .fromSTEB(
+                                                                            25.0,
+                                                                            5.0,
+                                                                            0.0,
+                                                                            0.0),
+                                                                child:
+                                                                    Container(
+                                                                  width: MediaQuery.sizeOf(
+                                                                              context)
+                                                                          .width *
+                                                                      0.45,
+                                                                  height: 100.0,
+                                                                  decoration:
+                                                                      BoxDecoration(
+                                                                    color: FlutterFlowTheme.of(
+                                                                            context)
+                                                                        .secondaryBackground,
+                                                                    borderRadius:
+                                                                        BorderRadius.circular(
+                                                                            7.0),
+                                                                    border:
+                                                                        Border
+                                                                            .all(
+                                                                      width:
+                                                                          1.0,
+                                                                    ),
+                                                                  ),
+                                                                  child:
+                                                                      ClipRRect(
+                                                                    borderRadius:
+                                                                        BorderRadius.circular(
+                                                                            8.0),
+                                                                    child: Image
+                                                                        .network(
+                                                                      valueOrDefault<
+                                                                          String>(
+                                                                        _model
+                                                                            .uploadedFileUrl,
+                                                                        'https://storage.googleapis.com/proudcity/mebanenc/uploads/2021/03/placeholder-image.png',
+                                                                      ),
+                                                                      width:
+                                                                          300.0,
+                                                                      height:
+                                                                          200.0,
+                                                                      fit: BoxFit
+                                                                          .cover,
+                                                                    ),
+                                                                  ),
                                                                 ),
-                                                        validator: _model
-                                                            .signimageControllerValidator
-                                                            .asValidator(
-                                                                context),
-                                                      ),
+                                                              ),
+                                                            ),
+                                                            Padding(
+                                                              padding:
+                                                                  const EdgeInsetsDirectional
+                                                                      .fromSTEB(
+                                                                          0.0,
+                                                                          0.0,
+                                                                          25.0,
+                                                                          0.0),
+                                                              child:
+                                                                  FFButtonWidget(
+                                                                onPressed:
+                                                                    () async {
+                                                                  final selectedMedia =
+                                                                      await selectMediaWithSourceBottomSheet(
+                                                                    context:
+                                                                        context,
+                                                                    allowPhoto:
+                                                                        true,
+                                                                  );
+                                                                  if (selectedMedia !=
+                                                                          null &&
+                                                                      selectedMedia.every((m) => validateFileFormat(
+                                                                          m.storagePath,
+                                                                          context))) {
+                                                                    setState(() =>
+                                                                        _model.isDataUploading =
+                                                                            true);
+                                                                    var selectedUploadedFiles =
+                                                                        <FFUploadedFile>[];
+
+                                                                    var downloadUrls =
+                                                                        <String>[];
+                                                                    try {
+                                                                      showUploadMessage(
+                                                                        context,
+                                                                        'Uploading file...',
+                                                                        showLoading:
+                                                                            true,
+                                                                      );
+                                                                      selectedUploadedFiles = selectedMedia
+                                                                          .map((m) => FFUploadedFile(
+                                                                                name: m.storagePath.split('/').last,
+                                                                                bytes: m.bytes,
+                                                                                height: m.dimensions?.height,
+                                                                                width: m.dimensions?.width,
+                                                                                blurHash: m.blurHash,
+                                                                              ))
+                                                                          .toList();
+
+                                                                      downloadUrls = (await Future
+                                                                              .wait(
+                                                                        selectedMedia
+                                                                            .map(
+                                                                          (m) async => await uploadData(
+                                                                              m.storagePath,
+                                                                              m.bytes),
+                                                                        ),
+                                                                      ))
+                                                                          .where((u) =>
+                                                                              u !=
+                                                                              null)
+                                                                          .map((u) =>
+                                                                              u!)
+                                                                          .toList();
+                                                                    } finally {
+                                                                      ScaffoldMessenger.of(
+                                                                              context)
+                                                                          .hideCurrentSnackBar();
+                                                                      _model.isDataUploading =
+                                                                          false;
+                                                                    }
+                                                                    if (selectedUploadedFiles.length ==
+                                                                            selectedMedia
+                                                                                .length &&
+                                                                        downloadUrls.length ==
+                                                                            selectedMedia.length) {
+                                                                      setState(
+                                                                          () {
+                                                                        _model.uploadedLocalFile =
+                                                                            selectedUploadedFiles.first;
+                                                                        _model.uploadedFileUrl =
+                                                                            downloadUrls.first;
+                                                                      });
+                                                                      showUploadMessage(
+                                                                          context,
+                                                                          'Success!');
+                                                                    } else {
+                                                                      setState(
+                                                                          () {});
+                                                                      showUploadMessage(
+                                                                          context,
+                                                                          'Failed to upload data');
+                                                                      return;
+                                                                    }
+                                                                  }
+                                                                },
+                                                                text: 'Upload',
+                                                                icon: const Icon(
+                                                                  Icons
+                                                                      .upload_file,
+                                                                  size: 15.0,
+                                                                ),
+                                                                options:
+                                                                    FFButtonOptions(
+                                                                  width: MediaQuery.sizeOf(
+                                                                              context)
+                                                                          .width *
+                                                                      0.35,
+                                                                  height: 40.0,
+                                                                  padding: const EdgeInsetsDirectional
+                                                                      .fromSTEB(
+                                                                          24.0,
+                                                                          0.0,
+                                                                          24.0,
+                                                                          0.0),
+                                                                  iconPadding: const EdgeInsetsDirectional
+                                                                      .fromSTEB(
+                                                                          0.0,
+                                                                          0.0,
+                                                                          0.0,
+                                                                          0.0),
+                                                                  color: FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .error,
+                                                                  textStyle: FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .titleSmall
+                                                                      .override(
+                                                                        fontFamily:
+                                                                            FlutterFlowTheme.of(context).titleSmallFamily,
+                                                                        color: Colors
+                                                                            .white,
+                                                                        useGoogleFonts:
+                                                                            GoogleFonts.asMap().containsKey(FlutterFlowTheme.of(context).titleSmallFamily),
+                                                                      ),
+                                                                  elevation:
+                                                                      3.0,
+                                                                  borderSide:
+                                                                      const BorderSide(
+                                                                    color: Colors
+                                                                        .transparent,
+                                                                    width: 1.0,
+                                                                  ),
+                                                                  borderRadius:
+                                                                      BorderRadius
+                                                                          .circular(
+                                                                              0.0),
+                                                                ),
+                                                              ),
+                                                            ),
+                                                          ],
+                                                        ),
+                                                      ],
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   24.0,
                                                                   12.0,
@@ -962,6 +1123,8 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                       child: TextFormField(
                                                         controller: _model
                                                             .signemailaddController,
+                                                        focusNode: _model
+                                                            .signemailaddFocusNode,
                                                         obscureText: false,
                                                         decoration:
                                                             InputDecoration(
@@ -988,7 +1151,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           enabledBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0xFFD1D1D1),
                                                               width: 1.0,
@@ -1001,7 +1164,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           focusedBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0x00000000),
                                                               width: 1.0,
@@ -1044,7 +1207,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                   .of(context)
                                                               .primaryBackground,
                                                           contentPadding:
-                                                              EdgeInsetsDirectional
+                                                              const EdgeInsetsDirectional
                                                                   .fromSTEB(
                                                                       20.0,
                                                                       24.0,
@@ -1075,7 +1238,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   24.0,
                                                                   12.0,
@@ -1084,6 +1247,8 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                       child: TextFormField(
                                                         controller: _model
                                                             .signnameController,
+                                                        focusNode: _model
+                                                            .signnameFocusNode,
                                                         obscureText: false,
                                                         decoration:
                                                             InputDecoration(
@@ -1110,7 +1275,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           enabledBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0xFFD1D1D1),
                                                               width: 1.0,
@@ -1123,7 +1288,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           focusedBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0x00000000),
                                                               width: 1.0,
@@ -1166,7 +1331,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                   .of(context)
                                                               .primaryBackground,
                                                           contentPadding:
-                                                              EdgeInsetsDirectional
+                                                              const EdgeInsetsDirectional
                                                                   .fromSTEB(
                                                                       20.0,
                                                                       24.0,
@@ -1197,7 +1362,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   24.0,
                                                                   12.0,
@@ -1206,6 +1371,8 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                       child: TextFormField(
                                                         controller: _model
                                                             .signpassController,
+                                                        focusNode: _model
+                                                            .signpassFocusNode,
                                                         obscureText: !_model
                                                             .signpassVisibility,
                                                         decoration:
@@ -1233,7 +1400,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           enabledBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0xFFD1D1D1),
                                                               width: 1.0,
@@ -1246,7 +1413,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           focusedBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0x00000000),
                                                               width: 1.0,
@@ -1289,7 +1456,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                   .of(context)
                                                               .primaryBackground,
                                                           contentPadding:
-                                                              EdgeInsetsDirectional
+                                                              const EdgeInsetsDirectional
                                                                   .fromSTEB(
                                                                       20.0,
                                                                       24.0,
@@ -1343,7 +1510,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   24.0,
                                                                   12.0,
@@ -1352,6 +1519,8 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                       child: TextFormField(
                                                         controller: _model
                                                             .signpassCheckController,
+                                                        focusNode: _model
+                                                            .signpassCheckFocusNode,
                                                         obscureText: !_model
                                                             .signpassCheckVisibility,
                                                         decoration:
@@ -1379,7 +1548,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           enabledBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0xFFD1D1D1),
                                                               width: 1.0,
@@ -1392,7 +1561,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           focusedBorder:
                                                               OutlineInputBorder(
                                                             borderSide:
-                                                                BorderSide(
+                                                                const BorderSide(
                                                               color: Color(
                                                                   0x00000000),
                                                               width: 1.0,
@@ -1435,7 +1604,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                   .of(context)
                                                               .primaryBackground,
                                                           contentPadding:
-                                                              EdgeInsetsDirectional
+                                                              const EdgeInsetsDirectional
                                                                   .fromSTEB(
                                                                       20.0,
                                                                       24.0,
@@ -1489,7 +1658,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   0.0,
                                                                   16.0,
@@ -1508,7 +1677,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                             ScaffoldMessenger
                                                                     .of(context)
                                                                 .showSnackBar(
-                                                              SnackBar(
+                                                              const SnackBar(
                                                                 content: Text(
                                                                   'Passwords don\'t match!',
                                                                 ),
@@ -1532,24 +1701,6 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                             return;
                                                           }
 
-                                                          await UserRecord
-                                                              .collection
-                                                              .doc(user.uid)
-                                                              .update(
-                                                                  createUserRecordData(
-                                                                email: _model
-                                                                    .signemailaddController
-                                                                    .text,
-                                                                displayName: _model
-                                                                    .signnameController
-                                                                    .text,
-                                                                photoUrl: _model
-                                                                    .signimageController
-                                                                    .text,
-                                                                createdTime:
-                                                                    getCurrentTimestamp,
-                                                              ));
-
                                                           context.goNamedAuth(
                                                               'HomePage',
                                                               context.mounted);
@@ -1560,14 +1711,14 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           width: 190.0,
                                                           height: 45.0,
                                                           padding:
-                                                              EdgeInsetsDirectional
+                                                              const EdgeInsetsDirectional
                                                                   .fromSTEB(
                                                                       0.0,
                                                                       0.0,
                                                                       0.0,
                                                                       0.0),
                                                           iconPadding:
-                                                              EdgeInsetsDirectional
+                                                              const EdgeInsetsDirectional
                                                                   .fromSTEB(
                                                                       0.0,
                                                                       0.0,
@@ -1575,7 +1726,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                       0.0),
                                                           color: FlutterFlowTheme
                                                                   .of(context)
-                                                              .mainC,
+                                                              .main1,
                                                           textStyle:
                                                               FlutterFlowTheme.of(
                                                                       context)
@@ -1593,7 +1744,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                   ),
                                                           elevation: 3.0,
                                                           borderSide:
-                                                              BorderSide(
+                                                              const BorderSide(
                                                             color: Colors
                                                                 .transparent,
                                                             width: 1.0,
@@ -1607,7 +1758,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   0.0,
                                                                   20.0,
@@ -1623,7 +1774,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                                   fontFamily: FlutterFlowTheme.of(
                                                                           context)
                                                                       .bodySmallFamily,
-                                                                  color: Color(
+                                                                  color: const Color(
                                                                       0xC5000000),
                                                                   fontSize:
                                                                       14.0,
@@ -1637,7 +1788,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                     ),
                                                     Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   0.0,
                                                                   4.0,
@@ -1652,7 +1803,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                         children: [
                                                           Padding(
                                                             padding:
-                                                                EdgeInsetsDirectional
+                                                                const EdgeInsetsDirectional
                                                                     .fromSTEB(
                                                                         8.0,
                                                                         8.0,
@@ -1663,7 +1814,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                               borderColor:
                                                                   FlutterFlowTheme.of(
                                                                           context)
-                                                                      .mainC,
+                                                                      .main1,
                                                               borderRadius:
                                                                   12.0,
                                                               borderWidth: 1.0,
@@ -1684,7 +1835,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           ),
                                                           Padding(
                                                             padding:
-                                                                EdgeInsetsDirectional
+                                                                const EdgeInsetsDirectional
                                                                     .fromSTEB(
                                                                         8.0,
                                                                         8.0,
@@ -1695,7 +1846,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                               borderColor:
                                                                   FlutterFlowTheme.of(
                                                                           context)
-                                                                      .mainC,
+                                                                      .main1,
                                                               borderRadius:
                                                                   12.0,
                                                               borderWidth: 1.0,
@@ -1716,7 +1867,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                           ),
                                                           Padding(
                                                             padding:
-                                                                EdgeInsetsDirectional
+                                                                const EdgeInsetsDirectional
                                                                     .fromSTEB(
                                                                         8.0,
                                                                         8.0,
@@ -1727,7 +1878,7 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                                                               borderColor:
                                                                   FlutterFlowTheme.of(
                                                                           context)
-                                                                      .mainC,
+                                                                      .main1,
                                                               borderRadius:
                                                                   12.0,
                                                               borderWidth: 1.0,
@@ -1773,53 +1924,8 @@ class _LoginPageWidgetState extends State<LoginPageWidget>
                           Container(
                             width: 200.0,
                             height: 200.0,
-                            decoration: BoxDecoration(),
+                            decoration: const BoxDecoration(),
                           ),
-                        Padding(
-                          padding: EdgeInsetsDirectional.fromSTEB(
-                              0.0, 0.0, 0.0, 30.0),
-                          child: Row(
-                            mainAxisSize: MainAxisSize.max,
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: [
-                              Padding(
-                                padding: EdgeInsetsDirectional.fromSTEB(
-                                    0.0, 0.0, 15.0, 0.0),
-                                child: Container(
-                                  width: 27.0,
-                                  height: 27.0,
-                                  clipBehavior: Clip.antiAlias,
-                                  decoration: BoxDecoration(
-                                    shape: BoxShape.circle,
-                                  ),
-                                  child: Image.asset(
-                                    'assets/images/Android_Playstore_Logo.png',
-                                    fit: BoxFit.cover,
-                                  ),
-                                ),
-                              ),
-                              Align(
-                                alignment: AlignmentDirectional(1.00, 0.00),
-                                child: Text(
-                                  'Made by Hoca inc.',
-                                  style: FlutterFlowTheme.of(context)
-                                      .bodyMedium
-                                      .override(
-                                        fontFamily: FlutterFlowTheme.of(context)
-                                            .bodyMediumFamily,
-                                        color: Color(0xFF939393),
-                                        fontSize: 15.0,
-                                        fontWeight: FontWeight.normal,
-                                        useGoogleFonts: GoogleFonts.asMap()
-                                            .containsKey(
-                                                FlutterFlowTheme.of(context)
-                                                    .bodyMediumFamily),
-                                      ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ),
                       ],
                     ),
                   ),
