@@ -2,25 +2,23 @@ import '/auth/firebase_auth/auth_util.dart';
 import '/backend/backend.dart';
 import '/backend/firebase_storage/storage.dart';
 import '/components/failed_popup_widget.dart';
-import '/components/sucess_pop_up_widget.dart';
 import '/flutter_flow/flutter_flow_drop_down.dart';
 import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
-import '/flutter_flow/flutter_flow_widgets.dart';
 import '/flutter_flow/form_field_controller.dart';
 import '/flutter_flow/upload_data.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
+import '/pages/sucess_pop_up/sucess_pop_up_widget.dart';
 import 'package:easy_debounce/easy_debounce.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_spinkit/flutter_spinkit.dart';
+import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 import 'c_upload_page_model.dart';
 export 'c_upload_page_model.dart';
 
 class CUploadPageWidget extends StatefulWidget {
-  const CUploadPageWidget({Key? key}) : super(key: key);
+  const CUploadPageWidget({super.key});
 
   @override
   _CUploadPageWidgetState createState() => _CUploadPageWidgetState();
@@ -37,7 +35,11 @@ class _CUploadPageWidgetState extends State<CUploadPageWidget> {
     _model = createModel(context, () => CUploadPageModel());
 
     _model.textController1 ??= TextEditingController();
+    _model.textFieldFocusNode1 ??= FocusNode();
+
     _model.textController2 ??= TextEditingController();
+    _model.textFieldFocusNode2 ??= FocusNode();
+
     WidgetsBinding.instance.addPostFrameCallback((_) => setState(() {}));
   }
 
@@ -50,216 +52,155 @@ class _CUploadPageWidgetState extends State<CUploadPageWidget> {
 
   @override
   Widget build(BuildContext context) {
+    if (isiOS) {
+      SystemChrome.setSystemUIOverlayStyle(
+        SystemUiOverlayStyle(
+          statusBarBrightness: Theme.of(context).brightness,
+          systemStatusBarContrastEnforced: true,
+        ),
+      );
+    }
+
     context.watch<FFAppState>();
 
     return GestureDetector(
-      onTap: () => FocusScope.of(context).requestFocus(_model.unfocusNode),
+      onTap: () => _model.unfocusNode.canRequestFocus
+          ? FocusScope.of(context).requestFocus(_model.unfocusNode)
+          : FocusScope.of(context).unfocus(),
       child: Scaffold(
         key: scaffoldKey,
         backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
-        body: SafeArea(
-          top: true,
-          child: Column(
-            mainAxisSize: MainAxisSize.max,
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Container(
-                width: MediaQuery.sizeOf(context).width * 1.0,
-                height: 65.0,
-                decoration: BoxDecoration(
-                  color: FlutterFlowTheme.of(context).main1,
-                ),
-                child: Row(
-                  mainAxisSize: MainAxisSize.max,
-                  children: [
-                    Padding(
-                      padding:
-                          EdgeInsetsDirectional.fromSTEB(15.0, 0.0, 7.0, 0.0),
-                      child: InkWell(
-                        splashColor: Colors.transparent,
-                        focusColor: Colors.transparent,
-                        hoverColor: Colors.transparent,
-                        highlightColor: Colors.transparent,
-                        onTap: () async {
-                          context.safePop();
-                        },
-                        child: Icon(
-                          Icons.arrow_back_ios,
-                          color: FlutterFlowTheme.of(context).primaryText,
-                          size: 24.0,
-                        ),
-                      ),
-                    ),
-                    Align(
-                      alignment: AlignmentDirectional(0.00, 0.00),
-                      child: Padding(
-                        padding:
-                            EdgeInsetsDirectional.fromSTEB(77.0, 0.0, 0.0, 0.0),
-                        child: Text(
-                          'yanggi  po`st',
-                          style: FlutterFlowTheme.of(context)
-                              .titleMedium
-                              .override(
-                                fontFamily: FlutterFlowTheme.of(context)
-                                    .titleMediumFamily,
-                                color: FlutterFlowTheme.of(context).primaryText,
-                                useGoogleFonts: GoogleFonts.asMap().containsKey(
-                                    FlutterFlowTheme.of(context)
-                                        .titleMediumFamily),
-                              ),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
+        body: Column(
+          mainAxisSize: MainAxisSize.max,
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Container(
+              width: MediaQuery.sizeOf(context).width * 1.0,
+              height: 65.0,
+              decoration: BoxDecoration(
+                color: FlutterFlowTheme.of(context).main1,
               ),
-              Padding(
-                padding: EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 0.0),
-                child: Container(
-                  width: MediaQuery.sizeOf(context).width * 1.0,
-                  decoration: BoxDecoration(
-                    color: FlutterFlowTheme.of(context).secondaryBackground,
-                  ),
-                  child: Padding(
+              child: Row(
+                mainAxisSize: MainAxisSize.max,
+                children: [
+                  Padding(
                     padding:
-                        EdgeInsetsDirectional.fromSTEB(0.0, 20.0, 0.0, 0.0),
-                    child: FlutterFlowDropDown<String>(
-                      controller: _model.dropDownValueController ??=
-                          FormFieldController<String>(null),
-                      options: [
-                        'mahalla savollar',
-                        'mahalla incidents ',
-                        'mahalliy oshxonalar',
-                        'mahalla yangiliklari',
-                        'odatlar',
-                        'yo`qolgan/topilgan ',
-                        'mahalla fotoko`rgazmasi',
-                        'harkuni'
-                      ],
-                      onChanged: (val) =>
-                          setState(() => _model.dropDownValue = val),
-                      width: 300.0,
-                      height: 50.0,
-                      textStyle: FlutterFlowTheme.of(context)
-                          .bodyMedium
-                          .override(
-                            fontFamily:
-                                FlutterFlowTheme.of(context).bodyMediumFamily,
-                            fontSize: 14.0,
-                            fontWeight: FontWeight.w500,
-                            useGoogleFonts: GoogleFonts.asMap().containsKey(
-                                FlutterFlowTheme.of(context).bodyMediumFamily),
-                          ),
-                      hintText: 'Category',
-                      icon: Icon(
-                        Icons.keyboard_arrow_down_rounded,
-                        color: FlutterFlowTheme.of(context).secondaryText,
+                        const EdgeInsetsDirectional.fromSTEB(15.0, 0.0, 7.0, 0.0),
+                    child: InkWell(
+                      splashColor: Colors.transparent,
+                      focusColor: Colors.transparent,
+                      hoverColor: Colors.transparent,
+                      highlightColor: Colors.transparent,
+                      onTap: () async {
+                        context.safePop();
+                      },
+                      child: Icon(
+                        Icons.arrow_back_ios,
+                        color: FlutterFlowTheme.of(context).primaryText,
                         size: 24.0,
                       ),
-                      fillColor:
-                          FlutterFlowTheme.of(context).secondaryBackground,
-                      elevation: 2.0,
-                      borderColor: FlutterFlowTheme.of(context).backG1,
-                      borderWidth: 2.0,
-                      borderRadius: 8.0,
-                      margin:
-                          EdgeInsetsDirectional.fromSTEB(16.0, 4.0, 16.0, 4.0),
-                      hidesUnderline: true,
-                      isSearchable: false,
-                      isMultiSelect: false,
                     ),
                   ),
+                  Align(
+                    alignment: const AlignmentDirectional(0.00, 0.00),
+                    child: Padding(
+                      padding:
+                          const EdgeInsetsDirectional.fromSTEB(77.0, 0.0, 0.0, 0.0),
+                      child: Text(
+                        'yanggi  po`st',
+                        style: FlutterFlowTheme.of(context)
+                            .titleMedium
+                            .override(
+                              fontFamily: FlutterFlowTheme.of(context)
+                                  .titleMediumFamily,
+                              color: FlutterFlowTheme.of(context).primaryText,
+                              useGoogleFonts: GoogleFonts.asMap().containsKey(
+                                  FlutterFlowTheme.of(context)
+                                      .titleMediumFamily),
+                            ),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 0.0),
+              child: Container(
+                width: MediaQuery.sizeOf(context).width * 1.0,
+                decoration: BoxDecoration(
+                  color: FlutterFlowTheme.of(context).secondaryBackground,
                 ),
-              ),
-              Container(
-                width: double.infinity,
-                decoration: BoxDecoration(),
-              ),
-              Container(
-                width: double.infinity,
-                height: 50.0,
-                decoration: BoxDecoration(),
                 child: Padding(
-                  padding: EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 0.0),
-                  child: TextFormField(
-                    controller: _model.textController1,
-                    onChanged: (_) => EasyDebounce.debounce(
-                      '_model.textController1',
-                      Duration(milliseconds: 2000),
-                      () => setState(() {}),
+                  padding: const EdgeInsetsDirectional.fromSTEB(0.0, 20.0, 0.0, 0.0),
+                  child: FlutterFlowDropDown<String>(
+                    controller: _model.dropDownValueController ??=
+                        FormFieldController<String>(null),
+                    options: const [
+                      'mahalla savollar',
+                      'mahalla incidents ',
+                      'mahalliy oshxonalar',
+                      'mahalla yangiliklari',
+                      'odatlar',
+                      'yo`qolgan/topilgan ',
+                      'mahalla fotoko`rgazmasi',
+                      'harkuni'
+                    ],
+                    onChanged: (val) =>
+                        setState(() => _model.dropDownValue = val),
+                    width: 300.0,
+                    height: 50.0,
+                    textStyle: FlutterFlowTheme.of(context).bodyMedium.override(
+                          fontFamily:
+                              FlutterFlowTheme.of(context).bodyMediumFamily,
+                          fontSize: 14.0,
+                          fontWeight: FontWeight.w500,
+                          useGoogleFonts: GoogleFonts.asMap().containsKey(
+                              FlutterFlowTheme.of(context).bodyMediumFamily),
+                        ),
+                    hintText: 'Category',
+                    icon: Icon(
+                      Icons.keyboard_arrow_down_rounded,
+                      color: FlutterFlowTheme.of(context).secondaryText,
+                      size: 24.0,
                     ),
-                    autofocus: true,
-                    textCapitalization: TextCapitalization.sentences,
-                    obscureText: false,
-                    decoration: InputDecoration(
-                      hintText: 'Title for your post',
-                      hintStyle: FlutterFlowTheme.of(context).labelLarge,
-                      enabledBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: FlutterFlowTheme.of(context).backG1,
-                          width: 2.0,
-                        ),
-                        borderRadius: const BorderRadius.only(
-                          topLeft: Radius.circular(4.0),
-                          topRight: Radius.circular(4.0),
-                        ),
-                      ),
-                      focusedBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: FlutterFlowTheme.of(context).primary,
-                          width: 2.0,
-                        ),
-                        borderRadius: const BorderRadius.only(
-                          topLeft: Radius.circular(4.0),
-                          topRight: Radius.circular(4.0),
-                        ),
-                      ),
-                      errorBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: FlutterFlowTheme.of(context).error,
-                          width: 2.0,
-                        ),
-                        borderRadius: const BorderRadius.only(
-                          topLeft: Radius.circular(4.0),
-                          topRight: Radius.circular(4.0),
-                        ),
-                      ),
-                      focusedErrorBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: FlutterFlowTheme.of(context).error,
-                          width: 2.0,
-                        ),
-                        borderRadius: const BorderRadius.only(
-                          topLeft: Radius.circular(4.0),
-                          topRight: Radius.circular(4.0),
-                        ),
-                      ),
-                      contentPadding:
-                          EdgeInsetsDirectional.fromSTEB(16.0, 8.0, 16.0, 12.0),
-                    ),
-                    style: FlutterFlowTheme.of(context).bodyLarge,
-                    textAlign: TextAlign.start,
-                    maxLines: 7,
-                    cursorColor: FlutterFlowTheme.of(context).backG1,
-                    validator:
-                        _model.textController1Validator.asValidator(context),
+                    fillColor: FlutterFlowTheme.of(context).secondaryBackground,
+                    elevation: 2.0,
+                    borderColor: FlutterFlowTheme.of(context).backG1,
+                    borderWidth: 2.0,
+                    borderRadius: 8.0,
+                    margin:
+                        const EdgeInsetsDirectional.fromSTEB(16.0, 4.0, 16.0, 4.0),
+                    hidesUnderline: true,
+                    isSearchable: false,
+                    isMultiSelect: false,
                   ),
                 ),
               ),
-              Padding(
-                padding: EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 0.0),
+            ),
+            Container(
+              width: double.infinity,
+              decoration: const BoxDecoration(),
+            ),
+            Container(
+              width: double.infinity,
+              height: 50.0,
+              decoration: const BoxDecoration(),
+              child: Padding(
+                padding: const EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 0.0),
                 child: TextFormField(
-                  controller: _model.textController2,
+                  controller: _model.textController1,
+                  focusNode: _model.textFieldFocusNode1,
                   onChanged: (_) => EasyDebounce.debounce(
-                    '_model.textController2',
-                    Duration(milliseconds: 2000),
+                    '_model.textController1',
+                    const Duration(milliseconds: 2000),
                     () => setState(() {}),
                   ),
                   autofocus: true,
                   textCapitalization: TextCapitalization.sentences,
                   obscureText: false,
                   decoration: InputDecoration(
-                    hintText: 'What\'s happening?',
+                    hintText: 'Title for your post',
                     hintStyle: FlutterFlowTheme.of(context).labelLarge,
                     enabledBorder: OutlineInputBorder(
                       borderSide: BorderSide(
@@ -302,208 +243,274 @@ class _CUploadPageWidgetState extends State<CUploadPageWidget> {
                       ),
                     ),
                     contentPadding:
-                        EdgeInsetsDirectional.fromSTEB(16.0, 8.0, 16.0, 12.0),
+                        const EdgeInsetsDirectional.fromSTEB(16.0, 8.0, 16.0, 12.0),
                   ),
                   style: FlutterFlowTheme.of(context).bodyLarge,
                   textAlign: TextAlign.start,
                   maxLines: 7,
                   cursorColor: FlutterFlowTheme.of(context).backG1,
                   validator:
-                      _model.textController2Validator.asValidator(context),
+                      _model.textController1Validator.asValidator(context),
                 ),
               ),
-              Padding(
-                padding: EdgeInsetsDirectional.fromSTEB(7.0, 0.0, 0.0, 0.0),
-                child: Container(
-                  width: 100.0,
-                  height: 100.0,
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(10.0),
-                    border: Border.all(
-                      color: FlutterFlowTheme.of(context).sub1,
+            ),
+            Padding(
+              padding: const EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 0.0),
+              child: TextFormField(
+                controller: _model.textController2,
+                focusNode: _model.textFieldFocusNode2,
+                onChanged: (_) => EasyDebounce.debounce(
+                  '_model.textController2',
+                  const Duration(milliseconds: 2000),
+                  () => setState(() {}),
+                ),
+                autofocus: true,
+                textCapitalization: TextCapitalization.sentences,
+                obscureText: false,
+                decoration: InputDecoration(
+                  hintText: 'What\'s happening?',
+                  hintStyle: FlutterFlowTheme.of(context).labelLarge,
+                  enabledBorder: OutlineInputBorder(
+                    borderSide: BorderSide(
+                      color: FlutterFlowTheme.of(context).backG1,
                       width: 2.0,
                     ),
-                  ),
-                  child: ClipRRect(
-                    borderRadius: BorderRadius.circular(8.0),
-                    child: Image.network(
-                      valueOrDefault<String>(
-                        _model.uploadedFileUrl,
-                        'https://cdn.xxl.thumbs.canstockphoto.com/add-photo-line-icon-image-thumbnail-sign-vector-add-photo-line-icon-image-thumbnail-sign-picture-eps-vectors_csp92477230.jpg',
-                      ),
-                      width: MediaQuery.sizeOf(context).width * 0.1,
-                      height: MediaQuery.sizeOf(context).height * 0.1,
-                      fit: BoxFit.cover,
+                    borderRadius: const BorderRadius.only(
+                      topLeft: Radius.circular(4.0),
+                      topRight: Radius.circular(4.0),
                     ),
+                  ),
+                  focusedBorder: OutlineInputBorder(
+                    borderSide: BorderSide(
+                      color: FlutterFlowTheme.of(context).primary,
+                      width: 2.0,
+                    ),
+                    borderRadius: const BorderRadius.only(
+                      topLeft: Radius.circular(4.0),
+                      topRight: Radius.circular(4.0),
+                    ),
+                  ),
+                  errorBorder: OutlineInputBorder(
+                    borderSide: BorderSide(
+                      color: FlutterFlowTheme.of(context).error,
+                      width: 2.0,
+                    ),
+                    borderRadius: const BorderRadius.only(
+                      topLeft: Radius.circular(4.0),
+                      topRight: Radius.circular(4.0),
+                    ),
+                  ),
+                  focusedErrorBorder: OutlineInputBorder(
+                    borderSide: BorderSide(
+                      color: FlutterFlowTheme.of(context).error,
+                      width: 2.0,
+                    ),
+                    borderRadius: const BorderRadius.only(
+                      topLeft: Radius.circular(4.0),
+                      topRight: Radius.circular(4.0),
+                    ),
+                  ),
+                  contentPadding:
+                      const EdgeInsetsDirectional.fromSTEB(16.0, 8.0, 16.0, 12.0),
+                ),
+                style: FlutterFlowTheme.of(context).bodyLarge,
+                textAlign: TextAlign.start,
+                maxLines: 7,
+                cursorColor: FlutterFlowTheme.of(context).backG1,
+                validator: _model.textController2Validator.asValidator(context),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsetsDirectional.fromSTEB(7.0, 0.0, 0.0, 0.0),
+              child: Container(
+                width: 100.0,
+                height: 100.0,
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(10.0),
+                  border: Border.all(
+                    color: FlutterFlowTheme.of(context).sub1,
+                    width: 2.0,
+                  ),
+                ),
+                child: ClipRRect(
+                  borderRadius: BorderRadius.circular(8.0),
+                  child: Image.network(
+                    valueOrDefault<String>(
+                      _model.uploadedFileUrl,
+                      'https://cdn.xxl.thumbs.canstockphoto.com/add-photo-line-icon-image-thumbnail-sign-vector-add-photo-line-icon-image-thumbnail-sign-picture-eps-vectors_csp92477230.jpg',
+                    ),
+                    width: MediaQuery.sizeOf(context).width * 0.1,
+                    height: MediaQuery.sizeOf(context).height * 0.1,
+                    fit: BoxFit.cover,
                   ),
                 ),
               ),
-              Padding(
-                padding: EdgeInsetsDirectional.fromSTEB(16.0, 12.0, 16.0, 0.0),
-                child: Row(
-                  mainAxisSize: MainAxisSize.max,
-                  children: [
-                    Padding(
-                      padding:
-                          EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 12.0, 0.0),
-                      child: FlutterFlowIconButton(
-                        borderColor: Colors.transparent,
-                        borderRadius: 30.0,
-                        borderWidth: 1.0,
-                        buttonSize: 60.0,
-                        icon: Icon(
-                          Icons.add_photo_alternate_outlined,
-                          color: FlutterFlowTheme.of(context).primary,
-                          size: 32.0,
-                        ),
-                        showLoadingIndicator: true,
-                        onPressed: () async {
-                          final selectedMedia =
-                              await selectMediaWithSourceBottomSheet(
-                            context: context,
-                            allowPhoto: true,
-                          );
-                          if (selectedMedia != null &&
-                              selectedMedia.every((m) =>
-                                  validateFileFormat(m.storagePath, context))) {
-                            setState(() => _model.isDataUploading = true);
-                            var selectedUploadedFiles = <FFUploadedFile>[];
+            ),
+            Padding(
+              padding: const EdgeInsetsDirectional.fromSTEB(16.0, 12.0, 16.0, 0.0),
+              child: Row(
+                mainAxisSize: MainAxisSize.max,
+                children: [
+                  Padding(
+                    padding:
+                        const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 12.0, 0.0),
+                    child: FlutterFlowIconButton(
+                      borderColor: Colors.transparent,
+                      borderRadius: 30.0,
+                      borderWidth: 1.0,
+                      buttonSize: 60.0,
+                      icon: Icon(
+                        Icons.add_photo_alternate_outlined,
+                        color: FlutterFlowTheme.of(context).primary,
+                        size: 32.0,
+                      ),
+                      showLoadingIndicator: true,
+                      onPressed: () async {
+                        final selectedMedia =
+                            await selectMediaWithSourceBottomSheet(
+                          context: context,
+                          allowPhoto: true,
+                        );
+                        if (selectedMedia != null &&
+                            selectedMedia.every((m) =>
+                                validateFileFormat(m.storagePath, context))) {
+                          setState(() => _model.isDataUploading = true);
+                          var selectedUploadedFiles = <FFUploadedFile>[];
 
-                            var downloadUrls = <String>[];
-                            try {
-                              selectedUploadedFiles = selectedMedia
-                                  .map((m) => FFUploadedFile(
-                                        name: m.storagePath.split('/').last,
-                                        bytes: m.bytes,
-                                        height: m.dimensions?.height,
-                                        width: m.dimensions?.width,
-                                        blurHash: m.blurHash,
-                                      ))
-                                  .toList();
+                          var downloadUrls = <String>[];
+                          try {
+                            selectedUploadedFiles = selectedMedia
+                                .map((m) => FFUploadedFile(
+                                      name: m.storagePath.split('/').last,
+                                      bytes: m.bytes,
+                                      height: m.dimensions?.height,
+                                      width: m.dimensions?.width,
+                                      blurHash: m.blurHash,
+                                    ))
+                                .toList();
 
-                              downloadUrls = (await Future.wait(
-                                selectedMedia.map(
-                                  (m) async =>
-                                      await uploadData(m.storagePath, m.bytes),
-                                ),
-                              ))
-                                  .where((u) => u != null)
-                                  .map((u) => u!)
-                                  .toList();
-                            } finally {
-                              _model.isDataUploading = false;
-                            }
-                            if (selectedUploadedFiles.length ==
-                                    selectedMedia.length &&
-                                downloadUrls.length == selectedMedia.length) {
-                              setState(() {
-                                _model.uploadedLocalFile =
-                                    selectedUploadedFiles.first;
-                                _model.uploadedFileUrl = downloadUrls.first;
-                              });
-                            } else {
-                              setState(() {});
-                              return;
-                            }
+                            downloadUrls = (await Future.wait(
+                              selectedMedia.map(
+                                (m) async =>
+                                    await uploadData(m.storagePath, m.bytes),
+                              ),
+                            ))
+                                .where((u) => u != null)
+                                .map((u) => u!)
+                                .toList();
+                          } finally {
+                            _model.isDataUploading = false;
                           }
-                        },
-                      ),
+                          if (selectedUploadedFiles.length ==
+                                  selectedMedia.length &&
+                              downloadUrls.length == selectedMedia.length) {
+                            setState(() {
+                              _model.uploadedLocalFile =
+                                  selectedUploadedFiles.first;
+                              _model.uploadedFileUrl = downloadUrls.first;
+                            });
+                          } else {
+                            setState(() {});
+                            return;
+                          }
+                        }
+                      },
                     ),
-                  ],
-                ),
+                  ),
+                ],
               ),
-              Padding(
-                padding:
-                    EdgeInsetsDirectional.fromSTEB(20.0, 180.0, 20.0, 20.0),
-                child: InkWell(
-                  splashColor: Colors.transparent,
-                  focusColor: Colors.transparent,
-                  hoverColor: Colors.transparent,
-                  highlightColor: Colors.transparent,
-                  onTap: () async {
-                    if ((_model.textController2.text != null &&
-                            _model.textController2.text != '') &&
-                        (_model.textController1.text != null &&
-                            _model.textController1.text != '') &&
-                        (_model.isDataUploading != null)) {
-                      await CommunityTalkRecord.collection
-                          .doc()
-                          .set(createCommunityTalkRecordData(
-                            category: _model.dropDownValue,
-                            content: _model.textController2.text,
-                            like: 0,
-                            title: _model.textController1.text,
-                            uploadTime: getCurrentTimestamp,
-                            uploadImage: _model.uploadedFileUrl,
-                            uploadUser: currentUserReference,
-                          ));
-                      await showModalBottomSheet(
-                        isScrollControlled: true,
-                        backgroundColor: Colors.transparent,
-                        enableDrag: false,
-                        context: context,
-                        builder: (context) {
-                          return GestureDetector(
-                            onTap: () => FocusScope.of(context)
-                                .requestFocus(_model.unfocusNode),
-                            child: Padding(
-                              padding: MediaQuery.viewInsetsOf(context),
-                              child: Container(
-                                height: MediaQuery.sizeOf(context).height * 1.0,
-                                child: SucessPopUpWidget(),
-                              ),
+            ),
+            Padding(
+              padding: const EdgeInsetsDirectional.fromSTEB(20.0, 180.0, 20.0, 20.0),
+              child: InkWell(
+                splashColor: Colors.transparent,
+                focusColor: Colors.transparent,
+                hoverColor: Colors.transparent,
+                highlightColor: Colors.transparent,
+                onTap: () async {
+                  if ((_model.textController2.text != '') &&
+                      (_model.textController1.text != '') &&
+                      (_model.isDataUploading != null)) {
+                    await CommunityTalkRecord.collection
+                        .doc()
+                        .set(createCommunityTalkRecordData(
+                          category: _model.dropDownValue,
+                          content: _model.textController2.text,
+                          like: 0,
+                          title: _model.textController1.text,
+                          uploadTime: getCurrentTimestamp,
+                          uploadImage: _model.uploadedFileUrl,
+                          uploadUser: currentUserReference,
+                        ));
+                    await showModalBottomSheet(
+                      isScrollControlled: true,
+                      backgroundColor: Colors.transparent,
+                      enableDrag: false,
+                      context: context,
+                      builder: (context) {
+                        return GestureDetector(
+                          onTap: () => _model.unfocusNode.canRequestFocus
+                              ? FocusScope.of(context)
+                                  .requestFocus(_model.unfocusNode)
+                              : FocusScope.of(context).unfocus(),
+                          child: Padding(
+                            padding: MediaQuery.viewInsetsOf(context),
+                            child: SizedBox(
+                              height: MediaQuery.sizeOf(context).height * 1.0,
+                              child: const SucessPopUpWidget(),
                             ),
-                          );
-                        },
-                      ).then((value) => safeSetState(() {}));
-                    } else {
-                      await showModalBottomSheet(
-                        isScrollControlled: true,
-                        backgroundColor: Colors.transparent,
-                        enableDrag: false,
-                        context: context,
-                        builder: (context) {
-                          return GestureDetector(
-                            onTap: () => FocusScope.of(context)
-                                .requestFocus(_model.unfocusNode),
-                            child: Padding(
-                              padding: MediaQuery.viewInsetsOf(context),
-                              child: Container(
-                                height: MediaQuery.sizeOf(context).height * 1.0,
-                                child: FailedPopupWidget(),
-                              ),
+                          ),
+                        );
+                      },
+                    ).then((value) => safeSetState(() {}));
+                  } else {
+                    await showModalBottomSheet(
+                      isScrollControlled: true,
+                      backgroundColor: Colors.transparent,
+                      enableDrag: false,
+                      context: context,
+                      builder: (context) {
+                        return GestureDetector(
+                          onTap: () => _model.unfocusNode.canRequestFocus
+                              ? FocusScope.of(context)
+                                  .requestFocus(_model.unfocusNode)
+                              : FocusScope.of(context).unfocus(),
+                          child: Padding(
+                            padding: MediaQuery.viewInsetsOf(context),
+                            child: SizedBox(
+                              height: MediaQuery.sizeOf(context).height * 1.0,
+                              child: const FailedPopupWidget(),
                             ),
-                          );
-                        },
-                      ).then((value) => safeSetState(() {}));
-                    }
-                  },
-                  child: Container(
-                    width: MediaQuery.sizeOf(context).width * 1.0,
-                    height: 60.0,
-                    decoration: BoxDecoration(
-                      color: FlutterFlowTheme.of(context).main1,
-                      borderRadius: BorderRadius.circular(10.0),
-                    ),
-                    child: Align(
-                      alignment: AlignmentDirectional(0.00, 0.00),
-                      child: Text(
-                        'Post',
-                        style: FlutterFlowTheme.of(context).titleLarge.override(
-                              fontFamily:
-                                  FlutterFlowTheme.of(context).titleLargeFamily,
-                              color: Colors.white,
-                              useGoogleFonts: GoogleFonts.asMap().containsKey(
-                                  FlutterFlowTheme.of(context)
-                                      .titleLargeFamily),
-                            ),
-                      ),
+                          ),
+                        );
+                      },
+                    ).then((value) => safeSetState(() {}));
+                  }
+                },
+                child: Container(
+                  width: MediaQuery.sizeOf(context).width * 1.0,
+                  height: 60.0,
+                  decoration: BoxDecoration(
+                    color: FlutterFlowTheme.of(context).main1,
+                    borderRadius: BorderRadius.circular(10.0),
+                  ),
+                  child: Align(
+                    alignment: const AlignmentDirectional(0.00, 0.00),
+                    child: Text(
+                      'Post',
+                      style: FlutterFlowTheme.of(context).titleLarge.override(
+                            fontFamily:
+                                FlutterFlowTheme.of(context).titleLargeFamily,
+                            color: Colors.white,
+                            useGoogleFonts: GoogleFonts.asMap().containsKey(
+                                FlutterFlowTheme.of(context).titleLargeFamily),
+                          ),
                     ),
                   ),
                 ),
               ),
-            ],
-          ),
+            ),
+          ],
         ),
       ),
     );
