@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:from_css_color/from_css_color.dart';
+
 import '/flutter_flow/flutter_flow_util.dart';
 
 export 'package:collection/collection.dart' show ListEquality;
@@ -22,8 +23,8 @@ List<T>? getStructList<T>(
     value is! List
         ? null
         : value
-            .where((e) => e is Map<String, dynamic>)
-            .map((e) => structBuilder(e as Map<String, dynamic>))
+            .whereType<Map<String, dynamic>>()
+            .map((e) => structBuilder(e))
             .toList();
 
 Color? getSchemaColor(dynamic value) => value is String
