@@ -1,19 +1,8 @@
-import '/auth/firebase_auth/auth_util.dart';
-import '/backend/backend.dart';
-import '/backend/firebase_storage/storage.dart';
-import '/flutter_flow/flutter_flow_icon_button.dart';
-import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
-import '/flutter_flow/flutter_flow_widgets.dart';
-import '/flutter_flow/upload_data.dart';
-import 'package:cached_network_image/cached_network_image.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
+import 'edit_profile_widget.dart' show EditProfileWidget;
 import 'package:flutter/material.dart';
-import 'package:flutter_spinkit/flutter_spinkit.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:provider/provider.dart';
 
-class EditProfileModel extends FlutterFlowModel {
+class EditProfileModel extends FlutterFlowModel<EditProfileWidget> {
   ///  State fields for stateful widgets in this page.
 
   bool isDataUploading = false;
@@ -22,22 +11,32 @@ class EditProfileModel extends FlutterFlowModel {
   String uploadedFileUrl = '';
 
   // State field(s) for yourName widget.
+  FocusNode? yourNameFocusNode;
   TextEditingController? yourNameController;
   String? Function(BuildContext, String?)? yourNameControllerValidator;
   // State field(s) for city widget.
+  FocusNode? cityFocusNode1;
   TextEditingController? cityController1;
   String? Function(BuildContext, String?)? cityController1Validator;
   // State field(s) for city widget.
+  FocusNode? cityFocusNode2;
   TextEditingController? cityController2;
   String? Function(BuildContext, String?)? cityController2Validator;
 
   /// Initialization and disposal methods.
 
+  @override
   void initState(BuildContext context) {}
 
+  @override
   void dispose() {
+    yourNameFocusNode?.dispose();
     yourNameController?.dispose();
+
+    cityFocusNode1?.dispose();
     cityController1?.dispose();
+
+    cityFocusNode2?.dispose();
     cityController2?.dispose();
   }
 
