@@ -77,7 +77,7 @@ class _FlutterFlowChoiceChipsState extends State<FlutterFlowChoiceChips> {
   @override
   void initState() {
     super.initState();
-    choiceChipValues = List.from(widget.controller.initialValue ?? []);
+    choiceChipValues = selectedValues;
     if (!widget.initialized && choiceChipValues.isNotEmpty) {
       SchedulerBinding.instance.addPostFrameCallback(
         (_) {
@@ -88,9 +88,6 @@ class _FlutterFlowChoiceChipsState extends State<FlutterFlowChoiceChips> {
       );
     }
     changeSelectedValues.addListener(() {
-      if (!listEquals(choiceChipValues, selectedValues)) {
-        setState(() => choiceChipValues = List.from(selectedValues));
-      }
       if (widget.onChanged != null) {
         widget.onChanged!(selectedValues);
       }
@@ -107,54 +104,57 @@ class _FlutterFlowChoiceChipsState extends State<FlutterFlowChoiceChips> {
   Widget build(BuildContext context) {
     final children = widget.options.map<Widget>(
       (option) {
-        final selected = choiceChipValues.contains(option.label);
+        final selected = selectedValues.contains(option.label);
         final style =
             selected ? widget.selectedChipStyle : widget.unselectedChipStyle;
-        return ChoiceChip(
-          selected: selected,
-          onSelected: widget.onChanged != null
-              ? (isSelected) {
-                  if (isSelected) {
-                    widget.multiselect
-                        ? choiceChipValues.add(option.label)
-                        : choiceChipValues = [option.label];
-                    widget.controller.value = List.from(choiceChipValues);
-                    setState(() {});
-                  } else {
-                    if (widget.multiselect) {
-                      choiceChipValues.remove(option.label);
+        return Theme(
+          data: Theme.of(context).copyWith(canvasColor: Colors.transparent),
+          child: ChoiceChip(
+            selected: selected,
+            onSelected: widget.onChanged != null
+                ? (isSelected) {
+                    if (isSelected) {
+                      widget.multiselect
+                          ? choiceChipValues.add(option.label)
+                          : choiceChipValues = [option.label];
                       widget.controller.value = List.from(choiceChipValues);
                       setState(() {});
+                    } else {
+                      if (widget.multiselect) {
+                        choiceChipValues.remove(option.label);
+                        widget.controller.value = List.from(choiceChipValues);
+                        setState(() {});
+                      }
                     }
                   }
-                }
-              : null,
-          label: Text(
-            option.label,
-            style: style.textStyle,
-          ),
-          labelPadding: style.labelPadding,
-          avatar: option.iconData != null
-              ? FaIcon(
-                  option.iconData,
-                  size: style.iconSize,
-                  color: style.iconColor,
-                )
-              : null,
-          elevation: style.elevation,
-          disabledColor: widget.disabledColor,
-          selectedColor:
-              selected ? widget.selectedChipStyle.backgroundColor : null,
-          backgroundColor:
-              selected ? null : widget.unselectedChipStyle.backgroundColor,
-          shape: RoundedRectangleBorder(
-            borderRadius: style.borderRadius ?? BorderRadius.circular(16),
-            side: BorderSide(
-              color: style.borderColor ?? Colors.transparent,
-              width: style.borderWidth ?? 0,
+                : null,
+            label: Text(
+              option.label,
+              style: style.textStyle,
             ),
+            labelPadding: style.labelPadding,
+            avatar: option.iconData != null
+                ? FaIcon(
+                    option.iconData,
+                    size: style.iconSize,
+                    color: style.iconColor,
+                  )
+                : null,
+            elevation: style.elevation,
+            disabledColor: widget.disabledColor,
+            selectedColor:
+                selected ? widget.selectedChipStyle.backgroundColor : null,
+            backgroundColor:
+                selected ? null : widget.unselectedChipStyle.backgroundColor,
+            shape: RoundedRectangleBorder(
+              borderRadius: style.borderRadius ?? BorderRadius.circular(16),
+              side: BorderSide(
+                color: style.borderColor ?? Colors.transparent,
+                width: style.borderWidth ?? 0,
+              ),
+            ),
+            materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
           ),
-          materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
         );
       },
     ).toList();
