@@ -7,10 +7,15 @@ export '/pages/home_page/home_page_widget.dart' show HomePageWidget;
 export '/pages/profile/profile_widget.dart' show ProfileWidget;
 export '/pages/post_detail_page/post_detail_page_widget.dart'
     show PostDetailPageWidget;
-export '/pages/search_page/search_page_widget.dart' show SearchPageWidget;
 export '/pages/community_page/community_page_widget.dart'
     show CommunityPageWidget;
 export '/pages/c_t_detail_page/c_t_detail_page_widget.dart'
     show CTDetailPageWidget;
 export '/pages/c_upload_page/c_upload_page_widget.dart' show CUploadPageWidget;
 export '/edit_profile/edit_profile_widget.dart' show EditProfileWidget;
+export '/pages/account/account_widget.dart' show AccountWidget;
+export '/nearby/nearby_widget.dart' show NearbyWidget;
+export '/nearby_detail_page/nearby_detail_page_widget.dart'
+    show NearbyDetailPageWidget;
+export '/f_a_qspage/f_a_qspage_widget.dart' show FAQspageWidget;
+export '/real_estate/real_estate_widget.dart' show RealEstateWidget;
