@@ -2,11 +2,10 @@ import '/backend/backend.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_toggle_icon.dart';
 import '/flutter_flow/flutter_flow_util.dart';
-import '/flutter_flow/flutter_flow_widgets.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart'
     as smooth_page_indicator;
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
@@ -15,9 +14,9 @@ export 'post_detail_page_model.dart';
 
 class PostDetailPageWidget extends StatefulWidget {
   const PostDetailPageWidget({
-    Key? key,
+    super.key,
     required this.postRef,
-  }) : super(key: key);
+  });
 
   final DocumentReference? postRef;
 
@@ -47,6 +46,15 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
 
   @override
   Widget build(BuildContext context) {
+    if (isiOS) {
+      SystemChrome.setSystemUIOverlayStyle(
+        SystemUiOverlayStyle(
+          statusBarBrightness: Theme.of(context).brightness,
+          systemStatusBarContrastEnforced: true,
+        ),
+      );
+    }
+
     context.watch<FFAppState>();
 
     return StreamBuilder<PostRecord>(
@@ -70,335 +78,333 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
         }
         final postDetailPagePostRecord = snapshot.data!;
         return GestureDetector(
-          onTap: () => FocusScope.of(context).requestFocus(_model.unfocusNode),
+          onTap: () => _model.unfocusNode.canRequestFocus
+              ? FocusScope.of(context).requestFocus(_model.unfocusNode)
+              : FocusScope.of(context).unfocus(),
           child: Scaffold(
             key: scaffoldKey,
             backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
-            body: SafeArea(
-              top: true,
+            body: Align(
+              alignment: const AlignmentDirectional(0.00, 0.00),
               child: Stack(
                 children: [
-                  SingleChildScrollView(
-                    child: Column(
-                      mainAxisSize: MainAxisSize.max,
-                      children: [
-                        Container(
-                          width: MediaQuery.sizeOf(context).width * 1.0,
-                          height: MediaQuery.sizeOf(context).height * 0.45,
-                          decoration: BoxDecoration(
-                            color: FlutterFlowTheme.of(context)
-                                .secondaryBackground,
-                          ),
-                          child: Stack(
-                            children: [
-                              Builder(
-                                builder: (context) {
-                                  final images = postDetailPagePostRecord
-                                      .uploadedImages
-                                      .toList();
-                                  return Container(
-                                    width: double.infinity,
-                                    height: 500.0,
-                                    child: Stack(
-                                      children: [
-                                        PageView.builder(
-                                          controller: _model
-                                                  .pageViewController ??=
-                                              PageController(
-                                                  initialPage: min(
-                                                      0, images.length - 1)),
-                                          scrollDirection: Axis.horizontal,
-                                          itemCount: images.length,
-                                          itemBuilder: (context, imagesIndex) {
-                                            final imagesItem =
-                                                images[imagesIndex];
-                                            return ClipRRect(
-                                              borderRadius:
-                                                  BorderRadius.circular(0.0),
-                                              child: Image.network(
-                                                imagesItem,
-                                                width: 300.0,
-                                                height: 200.0,
-                                                fit: BoxFit.cover,
-                                              ),
-                                            );
-                                          },
-                                        ),
-                                        Align(
-                                          alignment:
-                                              AlignmentDirectional(0.00, 1.00),
-                                          child: Padding(
-                                            padding:
-                                                EdgeInsetsDirectional.fromSTEB(
-                                                    0.0, 0.0, 0.0, 18.0),
-                                            child: smooth_page_indicator
-                                                .SmoothPageIndicator(
-                                              controller: _model
-                                                      .pageViewController ??=
-                                                  PageController(
-                                                      initialPage: min(0,
-                                                          images.length - 1)),
-                                              count: images.length,
-                                              axisDirection: Axis.horizontal,
-                                              onDotClicked: (i) async {
-                                                await _model.pageViewController!
-                                                    .animateToPage(
-                                                  i,
-                                                  duration: Duration(
-                                                      milliseconds: 500),
-                                                  curve: Curves.ease,
-                                                );
-                                              },
-                                              effect: smooth_page_indicator
-                                                  .SlideEffect(
-                                                spacing: 8.0,
-                                                radius: 16.0,
-                                                dotWidth: 8.0,
-                                                dotHeight: 8.0,
-                                                dotColor:
-                                                    FlutterFlowTheme.of(context)
-                                                        .accent1,
-                                                activeDotColor:
-                                                    FlutterFlowTheme.of(context)
-                                                        .primaryText,
-                                                paintStyle: PaintingStyle.fill,
-                                              ),
+                  Column(
+                    mainAxisSize: MainAxisSize.max,
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      Container(
+                        width: MediaQuery.sizeOf(context).width * 1.0,
+                        height: MediaQuery.sizeOf(context).height * 0.45,
+                        decoration: BoxDecoration(
+                          color:
+                              FlutterFlowTheme.of(context).secondaryBackground,
+                        ),
+                        child: Stack(
+                          children: [
+                            Builder(
+                              builder: (context) {
+                                final images = postDetailPagePostRecord
+                                    .uploadedImages
+                                    .toList();
+                                return SizedBox(
+                                  width: double.infinity,
+                                  height: 500.0,
+                                  child: Stack(
+                                    children: [
+                                      PageView.builder(
+                                        controller: _model
+                                                .pageViewController ??=
+                                            PageController(
+                                                initialPage:
+                                                    min(0, images.length - 1)),
+                                        scrollDirection: Axis.horizontal,
+                                        itemCount: images.length,
+                                        itemBuilder: (context, imagesIndex) {
+                                          final imagesItem =
+                                              images[imagesIndex];
+                                          return ClipRRect(
+                                            borderRadius:
+                                                BorderRadius.circular(0.0),
+                                            child: Image.network(
+                                              imagesItem,
+                                              width: 300.0,
+                                              height: 200.0,
+                                              fit: BoxFit.cover,
                                             ),
+                                          );
+                                        },
+                                      ),
+                                      Align(
+                                        alignment:
+                                            const AlignmentDirectional(0.00, 1.00),
+                                        child: Padding(
+                                          padding:
+                                              const EdgeInsetsDirectional.fromSTEB(
+                                                  0.0, 0.0, 0.0, 18.0),
+                                          child: smooth_page_indicator
+                                              .SmoothPageIndicator(
+                                            controller: _model
+                                                    .pageViewController ??=
+                                                PageController(
+                                                    initialPage: min(
+                                                        0, images.length - 1)),
+                                            count: images.length,
+                                            axisDirection: Axis.horizontal,
+                                            onDotClicked: (i) async {
+                                              await _model.pageViewController!
+                                                  .animateToPage(
+                                                i,
+                                                duration:
+                                                    const Duration(milliseconds: 500),
+                                                curve: Curves.ease,
+                                              );
+                                            },
+                                            effect: smooth_page_indicator
+                                                .SlideEffect(
+                                              spacing: 8.0,
+                                              radius: 16.0,
+                                              dotWidth: 8.0,
+                                              dotHeight: 8.0,
+                                              dotColor:
+                                                  FlutterFlowTheme.of(context)
+                                                      .accent1,
+                                              activeDotColor:
+                                                  FlutterFlowTheme.of(context)
+                                                      .primaryText,
+                                              paintStyle: PaintingStyle.fill,
+                                            ),
+                                          ),
+                                        ),
+                                      ),
+                                    ],
+                                  ),
+                                );
+                              },
+                            ),
+                            Container(
+                              width: MediaQuery.sizeOf(context).width * 1.0,
+                              height: MediaQuery.sizeOf(context).height * 0.08,
+                              decoration: const BoxDecoration(),
+                              child: Padding(
+                                padding: const EdgeInsetsDirectional.fromSTEB(
+                                    20.0, 0.0, 20.0, 0.0),
+                                child: Row(
+                                  mainAxisSize: MainAxisSize.max,
+                                  mainAxisAlignment:
+                                      MainAxisAlignment.spaceBetween,
+                                  children: [
+                                    Row(
+                                      mainAxisSize: MainAxisSize.min,
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.spaceBetween,
+                                      children: [
+                                        InkWell(
+                                          splashColor: Colors.transparent,
+                                          focusColor: Colors.transparent,
+                                          hoverColor: Colors.transparent,
+                                          highlightColor: Colors.transparent,
+                                          onTap: () async {
+                                            context.safePop();
+                                          },
+                                          child: const Icon(
+                                            Icons.arrow_back_ios,
+                                            color: Colors.white,
+                                            size: 24.0,
+                                          ),
+                                        ),
+                                        const Padding(
+                                          padding:
+                                              EdgeInsetsDirectional.fromSTEB(
+                                                  10.0, 0.0, 0.0, 0.0),
+                                          child: Icon(
+                                            Icons.cottage_sharp,
+                                            color: Colors.white,
+                                            size: 24.0,
                                           ),
                                         ),
                                       ],
                                     ),
-                                  );
-                                },
-                              ),
-                              Container(
-                                width: MediaQuery.sizeOf(context).width * 1.0,
-                                height:
-                                    MediaQuery.sizeOf(context).height * 0.08,
-                                decoration: BoxDecoration(),
-                                child: Padding(
-                                  padding: EdgeInsetsDirectional.fromSTEB(
-                                      20.0, 0.0, 20.0, 0.0),
-                                  child: Row(
-                                    mainAxisSize: MainAxisSize.max,
-                                    mainAxisAlignment:
-                                        MainAxisAlignment.spaceBetween,
-                                    children: [
-                                      Row(
-                                        mainAxisSize: MainAxisSize.min,
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.spaceBetween,
-                                        children: [
-                                          InkWell(
-                                            splashColor: Colors.transparent,
-                                            focusColor: Colors.transparent,
-                                            hoverColor: Colors.transparent,
-                                            highlightColor: Colors.transparent,
-                                            onTap: () async {
-                                              context.safePop();
-                                            },
-                                            child: Icon(
-                                              Icons.arrow_back_ios,
-                                              color: Colors.white,
-                                              size: 24.0,
-                                            ),
-                                          ),
-                                          Padding(
-                                            padding:
-                                                EdgeInsetsDirectional.fromSTEB(
-                                                    10.0, 0.0, 0.0, 0.0),
-                                            child: Icon(
-                                              Icons.cottage_sharp,
-                                              color: Colors.white,
-                                              size: 24.0,
-                                            ),
-                                          ),
-                                        ],
-                                      ),
-                                      Row(
-                                        mainAxisSize: MainAxisSize.min,
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.spaceBetween,
-                                        children: [
-                                          Padding(
-                                            padding:
-                                                EdgeInsetsDirectional.fromSTEB(
-                                                    0.0, 0.0, 10.0, 0.0),
-                                            child: Icon(
-                                              Icons.share_rounded,
-                                              color: Colors.white,
-                                              size: 24.0,
-                                            ),
-                                          ),
-                                          Icon(
-                                            Icons.more_vert,
+                                    const Row(
+                                      mainAxisSize: MainAxisSize.min,
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.spaceBetween,
+                                      children: [
+                                        Padding(
+                                          padding:
+                                              EdgeInsetsDirectional.fromSTEB(
+                                                  0.0, 0.0, 10.0, 0.0),
+                                          child: Icon(
+                                            Icons.share_rounded,
                                             color: Colors.white,
                                             size: 24.0,
                                           ),
-                                        ],
-                                      ),
-                                    ],
-                                  ),
+                                        ),
+                                        Icon(
+                                          Icons.more_vert,
+                                          color: Colors.white,
+                                          size: 24.0,
+                                        ),
+                                      ],
+                                    ),
+                                  ],
                                 ),
                               ),
-                            ],
-                          ),
+                            ),
+                          ],
                         ),
-                        StreamBuilder<UserRecord>(
-                          stream: UserRecord.getDocument(
-                              postDetailPagePostRecord.uploadUser!),
-                          builder: (context, snapshot) {
-                            // Customize what your widget looks like when it's loading.
-                            if (!snapshot.hasData) {
-                              return Center(
-                                child: SizedBox(
-                                  width: 50.0,
-                                  height: 50.0,
-                                  child: SpinKitChasingDots(
-                                    color: FlutterFlowTheme.of(context).info,
-                                    size: 50.0,
-                                  ),
+                      ),
+                      StreamBuilder<UserRecord>(
+                        stream: UserRecord.getDocument(
+                            postDetailPagePostRecord.uploadUser!),
+                        builder: (context, snapshot) {
+                          // Customize what your widget looks like when it's loading.
+                          if (!snapshot.hasData) {
+                            return Center(
+                              child: SizedBox(
+                                width: 50.0,
+                                height: 50.0,
+                                child: SpinKitChasingDots(
+                                  color: FlutterFlowTheme.of(context).info,
+                                  size: 50.0,
                                 ),
-                              );
-                            }
-                            final containerUserRecord = snapshot.data!;
-                            return Container(
-                              width: MediaQuery.sizeOf(context).width * 1.0,
-                              height: MediaQuery.sizeOf(context).height * 0.55,
-                              decoration: BoxDecoration(),
-                              child: Column(
-                                mainAxisSize: MainAxisSize.max,
-                                children: [
-                                  StreamBuilder<UserRecord>(
-                                    stream: UserRecord.getDocument(
-                                        postDetailPagePostRecord.uploadUser!),
-                                    builder: (context, snapshot) {
-                                      // Customize what your widget looks like when it's loading.
-                                      if (!snapshot.hasData) {
-                                        return Center(
-                                          child: SizedBox(
-                                            width: 50.0,
-                                            height: 50.0,
-                                            child: SpinKitChasingDots(
-                                              color:
-                                                  FlutterFlowTheme.of(context)
-                                                      .info,
-                                              size: 50.0,
-                                            ),
+                              ),
+                            );
+                          }
+                          final containerUserRecord = snapshot.data!;
+                          return Container(
+                            width: MediaQuery.sizeOf(context).width * 1.0,
+                            height: MediaQuery.sizeOf(context).height * 0.55,
+                            decoration: const BoxDecoration(),
+                            child: Column(
+                              mainAxisSize: MainAxisSize.max,
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [
+                                StreamBuilder<UserRecord>(
+                                  stream: UserRecord.getDocument(
+                                      postDetailPagePostRecord.uploadUser!),
+                                  builder: (context, snapshot) {
+                                    // Customize what your widget looks like when it's loading.
+                                    if (!snapshot.hasData) {
+                                      return Center(
+                                        child: SizedBox(
+                                          width: 50.0,
+                                          height: 50.0,
+                                          child: SpinKitChasingDots(
+                                            color: FlutterFlowTheme.of(context)
+                                                .info,
+                                            size: 50.0,
                                           ),
-                                        );
-                                      }
-                                      final containerUserRecord =
-                                          snapshot.data!;
-                                      return Container(
-                                        width:
-                                            MediaQuery.sizeOf(context).width *
-                                                1.0,
-                                        height:
-                                            MediaQuery.sizeOf(context).height *
-                                                0.1,
-                                        decoration: BoxDecoration(),
-                                        child: Row(
-                                          mainAxisSize: MainAxisSize.max,
-                                          children: [
-                                            Padding(
-                                              padding: EdgeInsetsDirectional
-                                                  .fromSTEB(
-                                                      15.0, 0.0, 0.0, 0.0),
-                                              child: Container(
-                                                width: 55.0,
-                                                height: 55.0,
-                                                clipBehavior: Clip.antiAlias,
-                                                decoration: BoxDecoration(
-                                                  shape: BoxShape.circle,
-                                                ),
-                                                child: Image.network(
-                                                  containerUserRecord.photoUrl,
-                                                  fit: BoxFit.cover,
-                                                ),
+                                        ),
+                                      );
+                                    }
+                                    final containerUserRecord = snapshot.data!;
+                                    return Container(
+                                      width: MediaQuery.sizeOf(context).width *
+                                          1.0,
+                                      height:
+                                          MediaQuery.sizeOf(context).height *
+                                              0.1,
+                                      decoration: const BoxDecoration(),
+                                      child: Row(
+                                        mainAxisSize: MainAxisSize.max,
+                                        children: [
+                                          Padding(
+                                            padding:
+                                                const EdgeInsetsDirectional.fromSTEB(
+                                                    15.0, 0.0, 0.0, 0.0),
+                                            child: Container(
+                                              width: 55.0,
+                                              height: 55.0,
+                                              clipBehavior: Clip.antiAlias,
+                                              decoration: const BoxDecoration(
+                                                shape: BoxShape.circle,
+                                              ),
+                                              child: Image.network(
+                                                containerUserRecord.photoUrl,
+                                                fit: BoxFit.cover,
                                               ),
                                             ),
-                                            Padding(
-                                              padding: EdgeInsetsDirectional
-                                                  .fromSTEB(
-                                                      10.0, 0.0, 0.0, 0.0),
-                                              child: Column(
-                                                mainAxisSize: MainAxisSize.max,
-                                                mainAxisAlignment:
-                                                    MainAxisAlignment.center,
-                                                crossAxisAlignment:
-                                                    CrossAxisAlignment.start,
-                                                children: [
-                                                  Padding(
-                                                    padding:
-                                                        EdgeInsetsDirectional
-                                                            .fromSTEB(0.0, 0.0,
-                                                                0.0, 10.0),
-                                                    child: Text(
-                                                      containerUserRecord
-                                                          .displayName,
-                                                      style:
-                                                          FlutterFlowTheme.of(
-                                                                  context)
-                                                              .labelSmall
-                                                              .override(
-                                                                fontFamily: FlutterFlowTheme.of(
-                                                                        context)
-                                                                    .labelSmallFamily,
-                                                                fontWeight:
-                                                                    FontWeight
-                                                                        .w600,
-                                                                useGoogleFonts: GoogleFonts
-                                                                        .asMap()
-                                                                    .containsKey(
-                                                                        FlutterFlowTheme.of(context)
-                                                                            .labelSmallFamily),
-                                                              ),
-                                                    ),
-                                                  ),
-                                                  Text(
-                                                    'something level',
+                                          ),
+                                          Padding(
+                                            padding:
+                                                const EdgeInsetsDirectional.fromSTEB(
+                                                    10.0, 0.0, 0.0, 0.0),
+                                            child: Column(
+                                              mainAxisSize: MainAxisSize.max,
+                                              mainAxisAlignment:
+                                                  MainAxisAlignment.center,
+                                              crossAxisAlignment:
+                                                  CrossAxisAlignment.start,
+                                              children: [
+                                                Padding(
+                                                  padding: const EdgeInsetsDirectional
+                                                      .fromSTEB(
+                                                          0.0, 0.0, 0.0, 10.0),
+                                                  child: Text(
+                                                    containerUserRecord
+                                                        .displayName,
                                                     style: FlutterFlowTheme.of(
                                                             context)
-                                                        .bodyMedium
+                                                        .labelSmall
                                                         .override(
                                                           fontFamily:
                                                               FlutterFlowTheme.of(
                                                                       context)
-                                                                  .bodyMediumFamily,
+                                                                  .labelSmallFamily,
                                                           fontWeight:
-                                                              FontWeight.w300,
+                                                              FontWeight.w600,
                                                           useGoogleFonts: GoogleFonts
                                                                   .asMap()
                                                               .containsKey(
                                                                   FlutterFlowTheme.of(
                                                                           context)
-                                                                      .bodyMediumFamily),
+                                                                      .labelSmallFamily),
                                                         ),
                                                   ),
-                                                ],
-                                              ),
+                                                ),
+                                                Text(
+                                                  'something level',
+                                                  style: FlutterFlowTheme.of(
+                                                          context)
+                                                      .bodyMedium
+                                                      .override(
+                                                        fontFamily:
+                                                            FlutterFlowTheme.of(
+                                                                    context)
+                                                                .bodyMediumFamily,
+                                                        fontWeight:
+                                                            FontWeight.w300,
+                                                        useGoogleFonts: GoogleFonts
+                                                                .asMap()
+                                                            .containsKey(
+                                                                FlutterFlowTheme.of(
+                                                                        context)
+                                                                    .bodyMediumFamily),
+                                                      ),
+                                                ),
+                                              ],
                                             ),
-                                          ],
-                                        ),
-                                      );
-                                    },
-                                  ),
-                                  Container(
-                                    width:
-                                        MediaQuery.sizeOf(context).width * 1.0,
-                                    height: MediaQuery.sizeOf(context).height *
-                                        0.35,
-                                    decoration: BoxDecoration(),
-                                    child: Padding(
-                                      padding: EdgeInsetsDirectional.fromSTEB(
-                                          15.0, 5.0, 15.0, 0.0),
+                                          ),
+                                        ],
+                                      ),
+                                    );
+                                  },
+                                ),
+                                Container(
+                                  width: MediaQuery.sizeOf(context).width * 1.0,
+                                  height:
+                                      MediaQuery.sizeOf(context).height * 0.29,
+                                  decoration: const BoxDecoration(),
+                                  child: Padding(
+                                    padding: const EdgeInsetsDirectional.fromSTEB(
+                                        15.0, 0.0, 15.0, 0.0),
+                                    child: SingleChildScrollView(
                                       child: Column(
                                         mainAxisSize: MainAxisSize.max,
                                         crossAxisAlignment:
                                             CrossAxisAlignment.start,
                                         children: [
-                                          Divider(
+                                          const Divider(
                                             thickness: 1.0,
                                             color: Color(0xFFB9B0B0),
                                           ),
@@ -426,7 +432,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                 MainAxisAlignment.start,
                                             children: [
                                               Padding(
-                                                padding: EdgeInsetsDirectional
+                                                padding: const EdgeInsetsDirectional
                                                     .fromSTEB(
                                                         0.0, 0.0, 20.0, 0.0),
                                                 child: Text(
@@ -441,7 +447,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                                     context)
                                                                 .labelSmallFamily,
                                                         color:
-                                                            Color(0xFF7A7878),
+                                                            const Color(0xFF7A7878),
                                                         decoration:
                                                             TextDecoration
                                                                 .underline,
@@ -468,7 +474,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                                       context)
                                                                   .labelSmallFamily,
                                                           color:
-                                                              Color(0xFF7A7878),
+                                                              const Color(0xFF7A7878),
                                                           useGoogleFonts: GoogleFonts
                                                                   .asMap()
                                                               .containsKey(
@@ -481,7 +487,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                           ),
                                           Padding(
                                             padding:
-                                                EdgeInsetsDirectional.fromSTEB(
+                                                const EdgeInsetsDirectional.fromSTEB(
                                                     0.0, 25.0, 0.0, 0.0),
                                             child: Text(
                                               postDetailPagePostRecord.contents,
@@ -507,20 +513,25 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                       ),
                                     ),
                                   ),
-                                  Container(
+                                ),
+                                Align(
+                                  alignment: const AlignmentDirectional(0.00, 0.00),
+                                  child: Container(
                                     width:
                                         MediaQuery.sizeOf(context).width * 1.0,
                                     height: MediaQuery.sizeOf(context).height *
                                         0.08,
-                                    decoration: BoxDecoration(),
+                                    decoration: const BoxDecoration(),
                                     child: Row(
                                       mainAxisSize: MainAxisSize.max,
                                       mainAxisAlignment:
                                           MainAxisAlignment.spaceEvenly,
+                                      crossAxisAlignment:
+                                          CrossAxisAlignment.center,
                                       children: [
                                         Padding(
                                           padding:
-                                              EdgeInsetsDirectional.fromSTEB(
+                                              const EdgeInsetsDirectional.fromSTEB(
                                                   10.0, 0.0, 0.0, 0.0),
                                           child: ToggleIcon(
                                             onPressed: () async {
@@ -545,8 +556,8 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                             ),
                                           ),
                                         ),
-                                        SizedBox(
-                                          height: 100.0,
+                                        const SizedBox(
+                                          height: 40.0,
                                           child: VerticalDivider(
                                             thickness: 1.0,
                                             color: Color(0xFFB8B4B4),
@@ -576,7 +587,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                                     context)
                                                                 .titleSmallFamily,
                                                         color:
-                                                            Color(0xC2C0BCBC),
+                                                            const Color(0xC2C0BCBC),
                                                         useGoogleFonts: GoogleFonts
                                                                 .asMap()
                                                             .containsKey(
@@ -589,10 +600,10 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                         ),
                                         Align(
                                           alignment:
-                                              AlignmentDirectional(0.00, 0.00),
+                                              const AlignmentDirectional(0.00, 0.00),
                                           child: Padding(
                                             padding:
-                                                EdgeInsetsDirectional.fromSTEB(
+                                                const EdgeInsetsDirectional.fromSTEB(
                                                     100.0, 0.0, 0.0, 0.0),
                                             child: InkWell(
                                               splashColor: Colors.transparent,
@@ -617,7 +628,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                               },
                                               child: Container(
                                                 width: 80.0,
-                                                height: 50.0,
+                                                height: 40.0,
                                                 decoration: BoxDecoration(
                                                   color: FlutterFlowTheme.of(
                                                           context)
@@ -628,7 +639,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                 ),
                                                 child: Align(
                                                   alignment:
-                                                      AlignmentDirectional(
+                                                      const AlignmentDirectional(
                                                           0.00, 0.00),
                                                   child: Text(
                                                     'Chat',
@@ -657,13 +668,13 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                       ],
                                     ),
                                   ),
-                                ],
-                              ),
-                            );
-                          },
-                        ),
-                      ],
-                    ),
+                                ),
+                              ],
+                            ),
+                          );
+                        },
+                      ),
+                    ],
                   ),
                 ],
               ),
